+++ lib/useDesktopBrowser.jsx
@@ -1,12 +1,12 @@
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
 
-import isMobileBrowser from '../utils/isMobileBrowser'
+import isMobileOrTableBrowser from '../utils/isMobileOrTablet'
 
 export default () => {
   const { replace } = useRouter()
   useEffect(() => {
-    const onMobileBrowser = isMobileBrowser()
+    const onMobileBrowser = isMobileOrTableBrowser()
     if (onMobileBrowser) replace('/error/use-desktop-browser')
   }, [replace])
 }
